backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Limiter==2.8.0
Flask-SQLAlchemy==2.5.1
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
torch==2.9.0
transformers==4.36.0 
tokenizers==0.15.2
sqlalchemy==2.0.20
python-dotenv==1.0.0
gunicorn==20.1.0
